Ajout d'un long int à la variable num pour éviter un overflox pour INT-MIN

+++ function_integer.c
@@ -10,7 +10,7 @@
 
 int print_int(va_list arg)
 {
-int num = va_arg(arg, int);
+long int num = va_arg(arg, int);
 char buffer[12];
 int count = 0;
 int neg = 0;
